Se agrego ReservaHabitacion, ademas se cambiaron todas las clases para el mapeo para persistir en la base de datos, con sus etiquetas correspondientes

+++ Desarrollo/src/main/java/tp/desarrollo/dto/PersonaDTO.java
@@ -8,15 +8,12 @@ import java.time.LocalDate;
 import tp.desarrollo.clases.Direccion;
 import tp.desarrollo.enums.TipoDocumento;
 
-/**
- *
- * @author juanc
- */
+
 public class PersonaDTO {
     private String apellido;
     private String nombre;
     private TipoDocumento tipo_documento;
-    private int num_documento;
+    private long num_documento;
     private long cuit;
     private LocalDate fecha_nacimiento;
     private Direccion direccion;
@@ -28,7 +25,7 @@ public class PersonaDTO {
         this.tipo_documento = tipo_documento;
         this.num_documento = Integer.parseInt(num_documento);
     }
-    public PersonaDTO(String apellido, String nombre, TipoDocumento tipo_documento, int num_documento, long cuit, LocalDate fecha_nacimiento, Direccion direccion, String nacionalidad) {
+    public PersonaDTO(String apellido, String nombre, TipoDocumento tipo_documento, long num_documento, long cuit, LocalDate fecha_nacimiento, Direccion direccion, String nacionalidad) {
         this.apellido = apellido;
         this.nombre = nombre;
         this.tipo_documento = tipo_documento;
@@ -47,7 +44,7 @@ public class PersonaDTO {
     public TipoDocumento getTipo_documento() {
         return tipo_documento;
     }
-    public int getNum_documento() {
+    public long getNum_documento() {
         return num_documento;
     }
     public long getCuit() {
@@ -72,7 +69,7 @@ public class PersonaDTO {
     public void setTipo_documento(TipoDocumento tipo_documento) {
         this.tipo_documento = tipo_documento;
     }
-    public void setNum_documento(int num_documento) {
+    public void setNum_documento(long num_documento) {
         this.num_documento = num_documento;
     }
     public void setCuit(long cuit) {
